feat(time): add get time info

+++ feature-time/src/main/java/com/powakaz/nesttrack/feature_time/data/datasourse/remote/model/TimeBalanceResponseDto.kt
@@ -11,6 +11,6 @@ data class TimeBalanceResponseDto(
     @SerialName("with_user_name")
     val userNameWith: String,
     @SerialName("balance_hours")
-    val balanceHours: Float
+    val balanceHours: Double
 ) {
 }
